requirements.txt
PyGithub
pandas
pytest
vcrpy
